Move mutation files to separate directory
