+++ saleor/graphql/checkout/tests/mutations/test_checkout_billing_address_update.py
@@ -1,5 +1,5 @@
-from ...core.utils import to_global_id_or_none
-from ...tests.utils import get_graphql_content
+from ....core.utils import to_global_id_or_none
+from ....tests.utils import get_graphql_content
 
 MUTATION_CHECKOUT_BILLING_ADDRESS_UPDATE = """
     mutation checkoutBillingAddressUpdate(
